requirements.txt
fastapi==0.115.13
psycopg2==2.9.10
python-dotenv==1.1.0
uvicorn==0.34.3
